requirements.txt
python-dotenv
pyTelegramBotAPI
shazamio
yt-dlp
